Generate size_t def at comptime, small refactor
Refactored statement.rs to use new argconv functions.
Generate size_t definition at compile time using bindgen.
