+++ scylla-rust-wrapper/src/argconv.rs
@@ -1,4 +1,4 @@
-use crate::size_t;
+use crate::types::size_t;
 use std::ffi::CStr;
 use std::os::raw::c_char;
