+++ scylla-rust-wrapper/src/types.rs
@@ -1,5 +1,8 @@
 #![allow(non_camel_case_types)]
 
+// Definition for size_t (and possibly other types in the future)
+include!(concat!(env!("OUT_DIR"), "/basic_types.rs"));
+
 pub type cass_bool_t = ::std::os::raw::c_uint;
 pub type cass_float_t = f32;
 pub type cass_double_t = f64;
@@ -13,4 +16,3 @@ pub type cass_int64_t = i64;
 pub type cass_uint64_t = u64;
 pub type cass_byte_t = cass_uint8_t;
 pub type cass_duration_t = cass_uint64_t;
-pub type size_t = ::std::os::raw::c_ulong;
